Переместил экраны в папку ui и добавил отступы в ресурсы
