+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.PracticumAndroidDiploma"
         tools:targetApi="31">
         <activity
-            android:name=".RootActivity"
+            android:name=".ui.RootActivity"
             android:screenOrientation="portrait"
             android:exported="true">
             <intent-filter>
