+++ app/src/main/java/ru/practicum/android/diploma/ui/RootActivity.kt
@@ -1,7 +1,9 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import ru.practicum.android.diploma.BuildConfig
+import ru.practicum.android.diploma.R
 
 class RootActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
