+++ app/src/main/java/ru/practicum/android/diploma/ui/filter_settings/fragments/FilterParametersFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.filter_settings.fragments
+package ru.practicum.android.diploma.ui.filter_settings.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.filter_settings.view_models.SelectAreaViewModel
+import ru.practicum.android.diploma.ui.filter_settings.view_models.FilterParametersViewModel
 
-class SelectAreaFragment : Fragment() {
+class FilterParametersFragment : Fragment() {
 
     companion object {
-        fun newInstance() = SelectAreaFragment()
+        fun newInstance() = FilterParametersFragment()
     }
 
-    private val viewModel: SelectAreaViewModel by viewModel<SelectAreaViewModel>()
+    private val viewModel: FilterParametersViewModel by viewModel<FilterParametersViewModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,6 +28,6 @@ class SelectAreaFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_select_area, container, false)
+        return inflater.inflate(R.layout.fragment_filter_parameters, container, false)
     }
 }
