+++ app/src/main/java/ru/practicum/android/diploma/ui/filter_settings/fragments/SelectIndustriesFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.filter_settings.fragments
+package ru.practicum.android.diploma.ui.filter_settings.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.filter_settings.view_models.SelectIndustriesViewModel
+import ru.practicum.android.diploma.ui.filter_settings.view_models.SelectIndustriesViewModel
 
 class SelectIndustriesFragment : Fragment() {
