+++ app/src/main/java/ru/practicum/android/diploma/ui/filter_settings/view_models/FilterParametersViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.filter_settings.view_models
+package ru.practicum.android.diploma.ui.filter_settings.view_models
 
 import androidx.lifecycle.ViewModel
