+++ app/src/main/java/ru/practicum/android/diploma/ui/vacancy_details/fragments/VacancyDetailsFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.filter_settings.fragments
+package ru.practicum.android.diploma.ui.vacancy_details.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.filter_settings.view_models.SelectLocationViewModel
+import ru.practicum.android.diploma.ui.vacancy_details.view_models.VacancyDetailsViewModel
 
-class SelectLocationFragment : Fragment() {
+class VacancyDetailsFragment : Fragment() {
 
     companion object {
-        fun newInstance() = SelectLocationFragment()
+        fun newInstance() = VacancyDetailsFragment()
     }
 
-    private val viewModel: SelectLocationViewModel by viewModel<SelectLocationViewModel>()
+    private val viewModel: VacancyDetailsViewModel by viewModel<VacancyDetailsViewModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,6 +28,6 @@ class SelectLocationFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_select_location, container, false)
+        return inflater.inflate(R.layout.fragment_vacancy_details, container, false)
     }
 }
