+++ app/src/main/java/ru/practicum/android/diploma/ui/vacancy_details/view_models/VacancyDetailsViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.vacancy_details.view_models
+package ru.practicum.android.diploma.ui.vacancy_details.view_models
 
 import androidx.lifecycle.ViewModel
