+++ app/src/main/res/layout/activity_root.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?backgroundPrimaryColor"
-    tools:context=".RootActivity">
+    tools:context=".ui.RootActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/container_view"
