+++ app/src/main/res/layout/fragment_fav_vacancies.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fav_vacancies.fragments.FavVacanciesFragment">
+    tools:context=".ui.fav_vacancies.fragments.FavVacanciesFragment">
 
     <TextView
         android:layout_width="match_parent"
