+++ app/src/main/res/layout/fragment_search_vacancies.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".search_vacancies.fragments.SearchVacanciesFragment">
+    tools:context=".ui.search_vacancies.fragments.SearchVacanciesFragment">
 
     <TextView
         android:layout_width="match_parent"
