+++ app/src/main/res/layout/fragment_select_area.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".filter_settings.fragments.SelectAreaFragment">
+    tools:context=".ui.filter_settings.fragments.SelectAreaFragment">
 
     <TextView
         android:layout_width="match_parent"
