+++ app/src/main/res/layout/fragment_select_industries.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".filter_settings.fragments.SelectIndustriesFragment">
+    tools:context=".ui.filter_settings.fragments.SelectIndustriesFragment">
 
     <TextView
         android:layout_width="match_parent"
