+++ app/src/main/res/layout/fragment_select_location.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".filter_settings.fragments.SelectLocationFragment">
+    tools:context=".ui.filter_settings.fragments.SelectLocationFragment">
 
     <TextView
         android:layout_width="match_parent"
