+++ app/src/main/res/layout/fragment_vacancy_details.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vacancy_details.fragments.VacancyDetailsFragment">
+    tools:context=".ui.vacancy_details.fragments.VacancyDetailsFragment">
 
     <TextView
         android:layout_width="match_parent"
