+++ app/src/main/res/navigation/navigation_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/searchVacanciesFragment"
-        android:name="ru.practicum.android.diploma.search_vacancies.fragments.SearchVacanciesFragment"
+        android:name="ru.practicum.android.diploma.ui.search_vacancies.fragments.SearchVacanciesFragment"
         android:label="fragment_search_vacancies"
         tools:layout="@layout/fragment_search_vacancies" >
         <action
@@ -19,17 +19,17 @@
     </fragment>
     <fragment
         android:id="@+id/creatorsFragment"
-        android:name="ru.practicum.android.diploma.creators.fragments.CreatorsFragment"
+        android:name="ru.practicum.android.diploma.ui.creators_fragments.CreatorsFragment"
         android:label="fragment_creators"
         tools:layout="@layout/fragment_creators" />
     <fragment
         android:id="@+id/favVacanciesFragment"
-        android:name="ru.practicum.android.diploma.fav_vacancies.fragments.FavVacanciesFragment"
+        android:name="ru.practicum.android.diploma.ui.fav_vacancies.fragments.FavVacanciesFragment"
         android:label="fragment_fav_vacancies"
         tools:layout="@layout/fragment_fav_vacancies" />
     <fragment
         android:id="@+id/filterParametersFragment"
-        android:name="ru.practicum.android.diploma.filter_settings.fragments.FilterParametersFragment"
+        android:name="ru.practicum.android.diploma.ui.filter_settings.fragments.FilterParametersFragment"
         android:label="fragment_filter_parameters"
         tools:layout="@layout/fragment_filter_parameters" >
         <action
@@ -41,17 +41,17 @@
     </fragment>
     <fragment
         android:id="@+id/vacancyDetailsFragment"
-        android:name="ru.practicum.android.diploma.vacancy_details.fragments.VacancyDetailsFragment"
+        android:name="ru.practicum.android.diploma.ui.vacancy_details.fragments.VacancyDetailsFragment"
         android:label="fragment_vacancy_details"
         tools:layout="@layout/fragment_vacancy_details" />
     <fragment
         android:id="@+id/selectAreaFragment"
-        android:name="ru.practicum.android.diploma.filter_settings.fragments.SelectAreaFragment"
+        android:name="ru.practicum.android.diploma.ui.filter_settings.fragments.SelectAreaFragment"
         android:label="fragment_select_area"
         tools:layout="@layout/fragment_select_area" />
     <fragment
         android:id="@+id/selectLocationFragment"
-        android:name="ru.practicum.android.diploma.filter_settings.fragments.SelectLocationFragment"
+        android:name="ru.practicum.android.diploma.ui.filter_settings.fragments.SelectLocationFragment"
         android:label="fragment_select_location"
         tools:layout="@layout/fragment_select_location" >
         <action
@@ -60,7 +60,7 @@
     </fragment>
     <fragment
         android:id="@+id/selectIndustriesFragment"
-        android:name="ru.practicum.android.diploma.filter_settings.fragments.SelectIndustriesFragment"
+        android:name="ru.practicum.android.diploma.ui.filter_settings.fragments.SelectIndustriesFragment"
         android:label="fragment_select_industries"
         tools:layout="@layout/fragment_select_industries" />
 </navigation>
