+++ app/src/main/res/values/dimens.xml
@@ -5,5 +5,11 @@
     <dimen name="text_size_lg">22sp</dimen>
     <dimen name="text_size_xl">32sp</dimen>
 
+    <dimen name="padding8">8dp</dimen>
+    <dimen name="padding16">16dp</dimen>
+    <dimen name="padding24">24dp</dimen>
+
+
     <dimen name="toolbar_height">64dp</dimen>
+
 </resources>
